bug in new config editor implementation

+++ Code/bash/file/autoEdit/configEditor/run.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
+#__________________________________________________ handle flags
 protected=0
-
 opts=':p'
 while getopts ${opts} opt; do
 	case ${opt} in
@@ -14,6 +14,7 @@ while getopts ${opts} opt; do
 	esac
 done
 
+#__________________________________________________ initialize variables
 # get current directory of script
 here="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
@@ -27,21 +28,39 @@ begin_marker="${2}"
 end_marker="${3}"
 confFile="${4}"
 
-touch ${newConfFile}
-sudo cp ${confFile} ${oldConfFile}
-sudo chmod a=rw ${oldConfFile}
-
+#__________________________________________________ awk editing / file manipulation
 export begin_marker end_marker
-command sudo ${awk}\
- -f ${editConfig}\
- -v begin_marker="${begin_marker}"\
- -v end_marker="${end_marker}"\
- -v new_section="${new_section}"\
- <${oldConfFile} >${newConfFile}
 
-sudo mv -f ${newConfFile} ${confFile}
+touch ${newConfFile}
 
-sudo chmod gu-w-r ${confFile}
+# writeprotected files
+if [ ${protected} -eq 1 ]; then
+	sudo cp ${confFile} ${oldConfFile}
+	sudo chmod a=rw ${oldConfFile}
+
+	sudo ${awk}\
+	 -f ${editConfig}\
+	 -v begin_marker="${begin_marker}"\
+	 -v end_marker="${end_marker}"\
+	 -v new_section="${new_section}"\
+	 <${oldConfFile} >${newConfFile}
+
+	sudo mv -f ${newConfFile} ${confFile}
+	sudo chmod gu-w-r ${confFile}
+# Non-writeprotected files
+else
+	cp ${confFile} ${oldConfFile}
+	chmod a=rw ${oldConfFile}
+
+	${awk}\
+	 -f ${editConfig}\
+	 -v begin_marker="${begin_marker}"\
+	 -v end_marker="${end_marker}"\
+	 -v new_section="${new_section}"\
+	 <${oldConfFile} >${newConfFile}
+
+	mv -f ${newConfFile} ${confFile}
+	chmod gu-w-r ${confFile}
+fi
 
 rm ${oldConfFile}
-
